t/verify.py: Add verify test script
Add a test script for verify. The tests are based on verify options and
it mostly focuses on norandommap. This test script includes two checksum
list where the first one contains minimal set, whereas the second one
contains full set of checksum which can be run by specifying -c or
--complete. It seems reasonable to test all of the data directions but
testing all of the checksum methods may be too much.

Signed-off-by: Vincent Fu <vincent.fu@samsung.com>
Signed-off-by: Ankit Kumar <ankit.kumar@samsung.com>